move authInstance out
